refactor: split MateController into domain-specific controllers
- MateApplicationController 분리
- MateInteractionController 분리
