+++ src/main/java/com/tripmoa/mate/controller/MateController.java
@@ -22,9 +22,6 @@ import java.util.List;
 @RequestMapping("/api/mate")
 public class MateController {
     private final MateService mateService;
-    private final MateApplicationService applyService;
-    private final MateLikeService likeService;
-    private final PassedPostService passedPostService;
 
     // 전체 메이트 포스트 조회
     @GetMapping("/")
@@ -67,96 +64,6 @@ public class MateController {
         User user = userDetails.getUser();
         this.mateService.deletePostById(id, user);
         return ResponseEntity.noContent().build();
-    }
-
-    // 내가 받은 메이트 신청서 조회
-    @PreAuthorize("isAuthenticated()")
-    @GetMapping("/applications/received")
-    public ResponseEntity<List<ApplicationResponse>> getReceivedApplication(
-            @AuthenticationPrincipal CustomUserDetails userDetails
-    ) {
-        Long userId = userDetails.getUser().getId();
-        List<ApplicationResponse> applications = this.applyService.getReceivedApplication(userId);
-        return ResponseEntity.ok().body(applications);
-    }
-
-    // 내가 보낸 메이트 신청서 조회
-    @PreAuthorize("isAuthenticated()")
-    @GetMapping("/applications/sent")
-    public ResponseEntity<List<ApplicationResponse>> getSentApplication(
-            @AuthenticationPrincipal CustomUserDetails userDetails
-    ) {
-        Long userId = userDetails.getUser().getId();
-        List<ApplicationResponse> applications = this.applyService.getSentApplication(userId);
-        return ResponseEntity.ok().body(applications);
-    }
-
-
-    // 특정 메이트 신청
-    @PostMapping("/{id}/apply/applicant")
-    public ResponseEntity<ApplicationResponse> applicateMatePost(
-            @PathVariable Long id,
-            @AuthenticationPrincipal CustomUserDetails userDetails,
-            @RequestBody ApplicationRequest request
-    ) {
-        User applicant = userDetails.getUser();
-        ApplicationResponse apply = this.applyService.createApply(id, request, applicant);
-        return ResponseEntity.ok().body(apply);
-    }
-
-    // 메이트 신청 승인
-    @PutMapping("/applications/{applyId}/approve")
-    public ResponseEntity<?> approve(
-            @PathVariable Long applyId,
-            @AuthenticationPrincipal CustomUserDetails userDetails
-    ) {
-        User author = userDetails.getUser();
-        ApplicationResponse approveApply = this.applyService.approveApply(applyId, author);
-        return ResponseEntity.ok().body(approveApply);
-    }
-
-    // 메이트 신청 거절
-    @PutMapping("/applications/{applyId}/reject")
-    public ResponseEntity<?> reject(
-            @PathVariable Long applyId,
-            @AuthenticationPrincipal CustomUserDetails userDetails
-    ) {
-        User author = userDetails.getUser();
-        ApplicationResponse rejectApply = this.applyService.rejectApply(applyId, author);
-        return ResponseEntity.ok().body(rejectApply);
-    }
-
-    // 좋아요 조회
-    @GetMapping("/{postId}/like")
-    public ResponseEntity<Long> getLikeCount(@PathVariable Long postId) {
-        Long likeCount = likeService.getLikeCount(postId);
-        return ResponseEntity.ok().body(likeCount);
-    }
-
-    // 좋아요 토글
-    @PostMapping("/{postId}/like")
-    public ResponseEntity<LikeResponse> toggleLike(
-            @PathVariable Long postId,
-            @AuthenticationPrincipal CustomUserDetails userDetails
-    ) {
-        Long userId = userDetails.getUser().getId();
-        LikeResponse response = likeService.toggleLike(postId, userId);
-        return ResponseEntity.ok().body(response);
-    }
-
-    // passed 게시글
-    @PostMapping("/posts/{postId}/pass")
-    public ResponseEntity<Void> pass(@PathVariable Long postId,
-                                     @AuthenticationPrincipal CustomUserDetails user) {
-        passedPostService.pass(user.getUser().getId(), postId);
-        return ResponseEntity.ok().build();
-    }
-
-    @DeleteMapping("/posts/{postId}/pass")
-    public ResponseEntity<Void> unpass(@PathVariable Long postId,
-                                       @AuthenticationPrincipal CustomUserDetails user) {
-        passedPostService.unpass(user.getUser().getId(), postId);
-        return ResponseEntity.ok().build();
     }
 
     @GetMapping("/posts/passed")
